src/test/java/com/down2thewire/RouteIdentifierTest.java
package com.down2thewire;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import static org.junit.jupiter.api.Assertions.*;

class RouteIdentifierTest {

    private RouteIdentifier routeIdentifier;
    private BranchGeoModel geoModel;
    private UserRouteRequest routeRequest;

    @BeforeEach
    void setUp() {
        // Initialize the geoModel and routeRequest with mock data or mocks
        geoModel = new BranchGeoModel();
        routeRequest = new UserRouteRequest();

        // Initialize RouteIdentifier with the mock data
        routeIdentifier = new RouteIdentifier(geoModel, routeRequest);
    }

    @Test
    void testGetBestRoute() {
        // Setup test conditions here, possibly with mock responses

        // Call the method under test
        LinearRoute result = routeIdentifier.getBestRoute("car", "walking", "fastest", 5000);

        // Perform assertions to verify the result
        assertNotNull(result);
        // Add more detailed checks based on the expected behavior of the method
    }

    @Test
    void testGetAlternativeVertex() {
        // Setup test conditions here, possibly with mock responses

        // Call the method under test
        LinearRoute result = routeIdentifier.getAlternativeVertex("car", "bus", "cheapest", "VertexX");

        // Perform assertions to verify the result
        assertNotNull(result);
        // Add more detailed checks based on the expected behavior of the method
    }

    // You could also add more tests for edge cases, invalid inputs, etc.

}
